delete audio player from case1 scene cam, add new soundtrack object, replace ost, create mixer in audio assets
The old audio player was attached to the camera which could create some confusion. If applied, this commit replaces the soundtrack player with an object called soundtrack  that plays the new song with reverb and loops. It also creates an audio mixer to easily mix the sfx and volume. Additionally, in an options menu, the volume could be exposed so that the player can adjust the level of the soundtrack and sfx