replace main menu music, adjust fader script, fix audio click on scene change
updated mix on main menu music. Fader script now generic and placed on each slider in the menu. At least on my end, it saves its values in the editor, but not when starting from scene 0, for ease of working instead of setting volume constantly while testing. Audio click was due to multiple audio listeners, make sure there are no audio listeners on the cameras.